Add react-router for /events and /fellows pages
