+++ web/package.json
@@ -11,6 +11,7 @@
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "reset-css": "^5.0.1",
     "styled-components": "^5.1.1"
